Persistence3.6

+++ DueThis_JavaLogic/src/controller/AccountController.java
@@ -5,40 +5,32 @@ import java.util.UUID;
 
 import model.Application;
 import model.Student;
+import persistence.SQLiteInterface;
 
 public class AccountController {
-
     public boolean createAccount(String uname, String pword, String email,
             boolean experiencedStudent, int sun, int mon, int tues, int wed, int thurs, int fri,
             int sat) throws InvalidInputException {
         String error = "";
 
-        // Check if required fields are null
-        if (uname == null || uname.trim().length() == 0) {
-
+        // - Check if required fields are null
+        if (uname == null || uname.trim().length() == 0)
             error += "Username is required to create an account! ";
-        }
 
-        if (pword == null || pword.trim().length() == 0) {
-
+        if (pword == null || pword.trim().length() == 0)
             error += "Password is required to create an account! ";
-        }
 
-        if (email == null || email.trim().length() == 0) {
-
+        if (email == null || email.trim().length() == 0)
             error += "Email is required to create an account! ";
-        } else if (!email.contains("@")) {
-
+        else if (!email.contains("@"))
             error += "Email address is invalid! ";
-        }
 
-        // Check if username or email address is already in use
+        // - Check if username or email address is already in use
         Application manager = Application.getInstance();
-
         int numAccounts = manager.numberOfStudents();
 
+        // - TODO Could be done in persistence using integrity constraints
         for (int i = 0; i < numAccounts; i++) {
-
             if (manager.getStudent(i).getUsername().equals(uname.trim())) {
                 error += "Username is already in use! ";
             }
@@ -48,42 +40,44 @@ public class AccountController {
             }
         }
 
-        // Check if the availabilities are valid integers if the student wishes
-        // to be experienced
+        // - TODO Code duplication
         if (experiencedStudent) {
-
-            // Make sure hours between 0 and 24 inclusive
             if (sun < 0 || sun > 24) {
                 error += "Sunday hours must be between 0 and 24! ";
             }
+
             if (mon < 0 || mon > 24) {
                 error += "Monday hours must be between 0 and 24! ";
             }
+
             if (tues < 0 || tues > 24) {
                 error += "Tuesday hours must be between 0 and 24! ";
             }
+
             if (wed < 0 || wed > 24) {
                 error += "Wednesday hours must be between 0 and 24! ";
             }
+
             if (thurs < 0 || thurs > 24) {
                 error += "Thursday hours must be between 0 and 24! ";
             }
+
             if (fri < 0 || fri > 24) {
                 error += "Friday hours must be between 0 and 24! ";
             }
+
             if (sat < 0 || sat > 24) {
                 error += "Saturday hours must be between 0 and 24! ";
             }
         }
 
-        // If there is any errors, throw the error
         if (error.length() > 0) {
             throw new InvalidInputException(error);
         }
 
         String id = UUID.randomUUID().toString();
 
-        // If the student wasnt experienced set all the availabilities to zero
+        // - For a novice student availabilities must be nil
         if (!experiencedStudent) {
             sun = 0;
             mon = 0;
@@ -94,19 +88,23 @@ public class AccountController {
             sat = 0;
         }
 
-        // Create the student
-        @SuppressWarnings("unused")
         Student s = new Student(id, uname, pword, email, experiencedStudent, sun, mon, tues, wed,
                 thurs, fri, sat, manager);
-        return true;
 
+        // - Try to commit to the database
+        SQLiteInterface.ensureConnection();
+
+        boolean op = SQLiteInterface.insertIntoStudents(SQLiteInterface.getConnection(), s);
+        if (!op)
+            throw new InvalidInputException("Failed to commit to database");
+
+        return op;
     }
 
     public boolean deleteAccount(String uname, String pword) throws InvalidInputException {
         String error = "";
         Application manager = Application.getInstance();
 
-        // If no students then we can't delete
         if (!manager.hasStudents()) {
             error = "No students have been created";
             throw new InvalidInputException(error);
@@ -116,11 +114,18 @@ public class AccountController {
         List<Student> studentList = manager.getStudents();
         for (Student s : studentList) {
             if (s.getUsername().equals(uname)) {
-                // If the student is found, we make sure that the passwords
-                // match
                 if (s.getPassword().equals(pword)) {
+                    // - Try to commit to the database
+                    // - FIXME Cascade delete
+                    SQLiteInterface.ensureConnection();
+
+                    boolean op = SQLiteInterface.deleteStudents(SQLiteInterface.getConnection(), s);
+                    if (!op)
+                        throw new InvalidInputException("Failed to commit to database");
+
                     s.delete();
-                    return true;
+
+                    return op;
                 } else {
                     error = "Password entered does not match";
                     throw new InvalidInputException(error);
@@ -228,13 +233,12 @@ public class AccountController {
         a.setFridayAvailability(fri);
         a.setSaturdayAvailability(sat);
 
-        // - TODO Persistence
-        // - TODO Consider an instance 'c' instead of making one.
-        manager.removeStudent(a);
-        manager.addStudent(a);
+        // - Try to commit to the database
+        SQLiteInterface.ensureConnection();
 
-        DueThisController c = new DueThisController();
-        persistence.SQLiteIntegration.saveDB(c.getPersistenceFilename(), manager);
+        boolean op = SQLiteInterface.updateStudents(SQLiteInterface.getConnection(), a);
+        if (!op)
+            throw new InvalidInputException("Failed to commit to database");
 
         return a;
     }
